Moved some optimiser functions around for the alternate perk on armor 1.0 and for the minimum armour total filter.
