+++ src/app/loadout-builder/generated-sets/GeneratedSetItem.tsx
@@ -23,7 +23,7 @@ import { LoadoutBuilderAction } from '../loadoutBuilderReducer';
  */
 function identifyAltPerkChoicesForChosenStats(item: DimItem, chosenValues: number[]) {
   let altPerks: DimPlug[] = [];
-  generateMixesFromPerks(item, {}, (mix, plugs) => {
+  generateMixesFromPerks(item, (mix, plugs) => {
     if (plugs && mix.every((val, index) => val === chosenValues[index])) {
       altPerks = plugs;
       return false;
